+++ cmd/forklift/plt/pallets.go
@@ -924,7 +924,11 @@ func cachePltAction(versions Versions) cli.ActionFunc {
 		}
 
 		fmt.Println("Downloading pallets specified by the local pallet...")
-		changed, err := fcli.DownloadRequiredPallets(0, plt, workspace.GetPalletCachePath())
+		cache, err := workspace.GetPalletCache()
+		if err != nil {
+			return err
+		}
+		changed, err := fcli.DownloadRequiredPallets(0, plt, cache)
 		if err != nil {
 			return err
 		}